Run and assess a machine learning method via resampling

This function handles the entire process.

run(
  x,
  y,
  resp.type = c("gaussian", "mgaussian", "binomial", "multinomial", "poisson", "cox"),
  filter = FALSE,
  threshold.na = 0.5,
  threshold.variability = 0.25,
  na.omit = FALSE,
  filter.methods = c("na", "median", "sd"),
  resampling.method = "multi.random",
  n = 10,
  min.obs = 5,
  n.grid.points = 5,
  balance = FALSE,
  do.parallel = FALSE,
  num.cores = 2,
  restore = TRUE,
  learning = list(learning.method = c("glmnet", "lasso", "ridge", "svm", "tree"),
    screening.method = c("none", "default", "alternative"), screening.args = NULL),
  confidence = 0.95,
  importance.term = "coefficient",
  random.seed = NULL,
  tolerance = 0,
  log.path = NULL,
  verbose = TRUE,
  outdir = NULL,
  filename = NULL,
  keep.call = FALSE,
  ...
)

Arguments

x
the input matrix, where rows are observations and columns are variables.

y
the response variable. Its number of rows must match the number of rows of x.

resp.type
the response type

filter
logical, whether to filter the variables that shows little or no variations across observations before resampling

threshold.na
threshold used in na filtering

threshold.variability
threshold used in top variable selection

na.omit
whether to apply na.omit after filtering.

filter.methods
the filtering method(s) to be used. If more than one method is selected, the order of execution is as reported below:

na

Remove a variable if sum of NA elements is greater than or equal a given percentage defined by threshold.na. Default is 0.5 (i.e. select variables with less than 50% of NAs)

median

Remove a variable if its mean expression is less than the median of the expression matrix

sd

Select highly variable variables depending on threshold.variability. Default is 0.25 (i.e. select variables in top 25%)

resampling.method
The resampling method to use. Two options are available:

'multi.random'

multiple random validation strategy

'cv'

k-fold cross-validation

n
the number of folds for method = "cv", or the number of repeats for method = "multi.random" or method = "bootstrap".

min.obs
minimum number of observations per class if resp.type= c("binomial", "multinomial") or minimum number of observations if resp.type= "gaussian"

n.grid.points
the desired number of grid points (n.grid.points=5 by default).

balance
whether to balance the classes in the training set (or in each fold) for bionomial, multinomial or cox response types.

do.parallel
logical, whether to try to parallelise the train-and-test procedure

num.cores
number of cores to use for parallel execution

restore
logical, whether to try to restore a previously ran analysis if it stopped before the end (outdir should have been set in the previous analysis)

confidence
probability for the random interval of falling on the estimated average accuracy measure

importance.term
the term to use in the computation of the significance score. Two options ara available:

coefficient

whether to use the normalised coefficient of a feature (default), i.e. the coefficient of the feature divided by the maximum absolute value of the coefficients of the features for that model so as to have a value ranging \([-1,1]\)

presence

whether to use a binary variable (presence = 0 or presence = 1) multiplied by the sign of the coefficient

random.seed
(optional) a vector of integers used as initial seeds for the random number generation during each step of the training-and-testing procedure. Its length should match n.grid.points. If only one number is provided, the seed is set just once at the beginning. If no number is provided, one is automatically generated and set at the begining (default).

tolerance
numeric, it indicates whether to use a tolerance in the selection of the best model, meaning if a model with a lower number of variables should be selected even if loosing a bit of prediction power.

log.path
A path to the log file to use.

verbose
logical. Should R report extra information on progress? Set to TRUE

outdir
path to the output directory

filename
(optional) name without extension for the output file

keep.call
logical, whether to keep the function call

...
further arguments to learning.method

learning.method
the learning method to use

screening.method
the type of filtering to apply

default

if default is selected, the adopted test is different depending on the responste type:

gaussian

test is correlation

mgaussian

test is correlation (for each gene, one correlation is computed with each response variable)

binomial

test is t-test or Wilcox test, depending on the number of observations

multinomial

test is anova or Kruskall-Wallis test, depending on the number of observations

poisson

test is correlation

cox

test is a Cox proportional hazards regression model

alternative

alternative test uses compute_sam and compute_limma. It is slower than "default".

screening.args
further arguments to screening method

Value

An object of class renoir. It contains different elements:

x

the filtered input matrix, where each row is an observation vector

y

the filtered response variable

resp.type

the response type

learning.method

the used learning method

resampling.method

the used resampling method

screening.method

the used screening method

balanced

whether the classes in the training set were balanced

train.sizes

a vector containing the considered training set sizes

tat

(train-and-test) a list of renoir.tat objects, one for each training set size. See tat_model for more details.

ibest

the index of the best training set size, for each best configuration ('min' and '1se'), given each accuracy measure computed over the test, train and full sets of data

confidence

the confidence used in the computation of the confidence intervals for the mean measure of accuracy obtained from the assessment of the trained models over the assessment sets

set

the set of data used for the computation of the accuracy measures considered for the selection of the best training set size

tolerance

the tolerance used in the selection of the best model

features.importance

the features importance, for each best configuration ('min' and '1se'), given each accuracy measure computed over the test and the full sets

importance.term

the term used in the computation of the significance score

measures

a vector containing the names of the accuracy measures used to assess the trained models

filter.summary

a text summary of the adopted initial filtering. If no pre-processing was used, it is set to NA

random.seed

a vector containing the numbers used as initial seeds for the random number generation during each step of the training-and-testing procedure; or a number used as initial seed for the random number generation at the beginning of the run

Author

Alessandro Barberis