Pre-processing the input data

This function filters the input matrix x depending on 1) presence of NAs, 2) mean expression of variable > median of expression matrix, 3) selection of top variable features. See filter below for further details. If observations are removed from x then the response variable y is updated accordingly.

run_filtering(
  x,
  y,
  filter = TRUE,
  filter.methods = c("na", "median", "sd"),
  threshold.na = 0.5,
  threshold.variability = 0.25,
  na.omit = FALSE,
  verbose = TRUE,
  outdir,
  filename,
  log.path
)

Arguments

x
the input matrix, where rows are observations and columns are variables.

y
the response variable. Its number of rows must match the number of rows of x.

filter
logical, whether to filter the variables that shows little or no variations across observations before resampling

filter.methods
the filtering method(s) to be used. If more than one method is selected, the order of execution is as reported below:

na

Remove a variable if sum of NA elements is greater than or equal a given percentage defined by threshold.na. Default is 0.5 (i.e. select variables with less than 50% of NAs)

median

Remove a variable if its mean expression is less than the median of the expression matrix

sd

Select highly variable variables depending on threshold.variability. Default is 0.25 (i.e. select variables in top 25%)

threshold.na
threshold used in na filtering

threshold.variability
threshold used in top variable selection

na.omit
whether to apply na.omit after filtering.

verbose
logical. Should R report extra information on progress? Set to TRUE

outdir
path to the output directory

filename
name of the file without extension (e.g. filename = 'file1'), used to store the filtered matrix if outdir is provided. If outdir is provided and filename is missing, a file name is automatically generated ('filteredMatrix').

log.path
A path to the log file to use.

Value

A list containing three elements

x

the filtered input matrix, where each row is an observation vector.

y

the filtered response variable.

text.summary

a text summary of the filtering.

Author

Alessandro Barberis

Examples

# Gaussian
x=matrix(rnorm(100*20),100,20)
y=rnorm(100)

f = run_filtering(x = x,
                 y=y,
                 threshold.na = 0.5,
                 threshold.variability = 0.25,
                 filter = TRUE,
                 na.omit = FALSE,
                 log.path = NULL,
                 verbose = TRUE,
                 outdir = NULL)
#> Error in check.output.dir(outdir): could not find function "check.output.dir"