Run and assess a machine learning method

This function handles the entire process of running and assessing a machine learning method via a multiple random strategy.

run_multirandom(
  x,
  y,
  resp.type = c("gaussian", "mgaussian", "binomial", "multinomial", "poisson", "cox"),
  n = 10,
  min.obs = 5,
  n.grid.points = 5,
  balance = FALSE,
  do.parallel = FALSE,
  num.cores = 2,
  learning.method = "glmnet",
  screening.method = c("none", "default", "alternative"),
  screening.args = NULL,
  confidence = 0.95,
  random.seed = NULL,
  tolerance = 0,
  verbose = TRUE,
  outdir = NULL,
  log.path = NULL,
  keep.call = FALSE,
  restore = TRUE,
  ...
)

Arguments

x
the input matrix, where rows are observations and columns are variables.

y
the response variable. Its number of rows must match the number of rows of x.

resp.type
the response type

n
the number of folds for method = "cv", or the number of repeats for method = "multi.random" or method = "bootstrap".

min.obs
minimum number of observations per class if resp.type= c("binomial", "multinomial") or minimum number of observations if resp.type= "gaussian"

n.grid.points
the desired number of grid points (n.grid.points=5 by default).

balance
whether to balance the classes in the training set (or in each fold) for bionomial, multinomial or cox response types.

do.parallel
logical, whether to try to parallelise the train-and-test procedure

num.cores
number of cores to use for parallel execution

learning.method
the learning method to use

screening.method
the type of filtering to apply

default

if default is selected, the adopted test is different depending on the responste type:

gaussian

test is correlation

mgaussian

test is correlation (for each gene, one correlation is computed with each response variable)

binomial

test is t-test or Wilcox test, depending on the number of observations

multinomial

test is anova or Kruskall-Wallis test, depending on the number of observations

poisson

test is correlation

cox

test is a Cox proportional hazards regression model

alternative

alternative test uses compute_sam and compute_limma. It is slower than "default".

screening.args
further arguments to screening method

confidence
probability for the random interval of falling on the estimated average accuracy measure

random.seed
(optional) a vector of integers used as initial seeds for the random number generation during each step of the training-and-testing procedure. Its length should match n.grid.points. If only one number is provided, the seed is set just once at the beginning. If no number is provided, one is automatically generated and set at the begining (default).

tolerance
numeric, it indicates whether to use a tolerance in the selection of the best model, meaning if a model with a lower number of variables should be selected even if loosing a bit of prediction power.

verbose
logical. Should R report extra information on progress? Set to TRUE

outdir
path to the output directory

log.path
A path to the log file to use.

keep.call
logical, whether to keep the function call

restore
logical, whether to try to restore a previously ran analysis if it stopped before the end (outdir should have been set in the previous analysis). If TRUE, the function looks for the existance of a result file for each training set size. If the file exists, the function loads the results; if the result file is not found, the analysis is started by using the selected training set size.

...
further arguments to learning.method

Author

Alessandro Barberis