#'Train and Test #'@inheritParams run #'@param train.size training set size, used for "multi.random" #'@param n the number of times to resample given a training set size #'for "multi.random" and "bootstrap" methods, or the number of folds for "cv". #'@param balance whether to balance the classes in the training set (or in each fold) #'for bionomial or multinomial response types. #'@inheritParams which_best_resampling_iteration #'@param random.seed (optional) integer, initial seed for the random number generator #'@return An object of class renoir.tat with different elements #'
#'
resampling.method

the used resampling method

#'
random.seed

the number used as initial seed for the random number generation #' during the training-and-testing procedure

#'
n

the number of times of resampling

#'
train

a list of length n, where each element is an object of class renoir.trained

#'
itrain

a list of length n, where each element is a vector containing #' the observations indices for the resampling iteration

#'
assess

a list of length 3: #'
#'
train.set

a list of length 2. For each configuration, there is a list containing the #' confidence intervals and the performance measures computed over the training set: #'
#'
ci

a list of confidence intervals for the mean value of each performance measure

#'
measures

a list of performance measures, where each element is a vector of measures

#'
#'

#'
test.set

a list of length 2. For each configuration, there is a list containing the #' confidence intervals and the performance measures computed over the testing set: #'
#'
ci

a list of confidence intervals for the mean value of each performance measure

#'
measures

a list of performance measures, where each element is a vector of measures

#'
#'

#'
full.set

a list of length 2. For each configuration, there is a list containing the #' confidence intervals and the performance measures computed over the entire dataset: #'
#'
ci

a list of confidence intervals for the mean value of each performance measure

#'
measures

a list of performance measures, where each element is a vector of measures

#'
#'

#'
#'

#'
ibest

a list of length 3: #'
#'
train.set

for each configuration, there is a list of indices for the best model given each performance measure computed over the training set

#'
test.set

for each configuration, there is a list of indices for the best model given each performance measure computed over the testing set

#'
full.set

for each configuration, there is a list of indices for the best model given each performance measure computed over the full data set

#'
#'

#'
precruit

a list of 2 elements; for each best configuration, a data frame containing the probability of recruitment for each feature of the model #'
#'
best.config.min

a data frame containing the probability of recruitment for each feature of the model

#'
best.config.1se

a data frame containing the probability of recruitment for each feature of the model

#'
#'

#'
call

(optional) the arguments of the call that produced this object

#'
resample <- function(x, y, resp.type = c("gaussian","mgaussian", "binomial","multinomial" ,"poisson","cox"),

#Resample resampling.method = c("multi.random", "bootstrap", "cv"), train.size, balance = FALSE, n = 10, do.parallel = FALSE, num.cores = 2,

tat_model_multirandom(
  x,
  y,
  resp.type = c("gaussian", "mgaussian", "binomial", "multinomial", "poisson", "cox"),
  train.size,
  balance = FALSE,
  n = 10,
  do.parallel = FALSE,
  num.cores = 2,
  learning.method = "glmnet",
  screening.method = c("none", "default", "alternative"),
  screening.args = NULL,
  confidence = 0.95,
  tolerance = 0,
  random.seed = NULL,
  verbose = TRUE,
  outdir,
  filename,
  log.path = NULL,
  keep.call = FALSE,
  ...
)

Value

An object with different elements

train

a list of length n, where each element contains

itrain

the indices of observations used for training for the resampling iteration

train.res

a list with different components, containing the fitted model over the resampled training set. The specific components depends on the fit.method

test

a list of length n, where each element contains

assess.best.config.min

performance measures for the model given by the best configuration

assess.best.config.1se

performance measures for the model given by the best configuration (1 standard error rule)

assess

a list of length 2

best.config.min

a list of length 2

ci

a list of confidence intervals for the mean value of each performance measure

measures

a list of performance measures, where each element is a vector of measures

best.config.1se

a list of length 2

ci

a list of confidence intervals for the mean value of each performance measure

measures

a list of performance measures, where each element is a vector of measures

ibest

for each best configuration, there is a list of indices for the best model given each performance measure

best.config.min

a list where each element is associated to a performance measure

best.config.1se

a list where each element is associated to a performance measure

precruit

for each best configuration, a data frame containing the probability of recruitment for each feature of the model

best.config.min

a data frame containing the probability of recruitment for each feature of the model

best.config.1se

a data frame containing the probability of recruitment for each feature of the model

call

(optional) the arguments of the call that produced this object

Details

#Learning method learning.method = "glmnet",

#Screening screening.method = c("none", "default", "alternative"), screening.args = NULL,

#Confidence intervals confidence = 0.95,

#Best model tolerance = 0,

#Further args random.seed = 123, verbose = TRUE, outdir, filename, log.path, keep.call = FALSE, ...) #Store the arguments call args <- c(as.list(environment()), list(...)); if(keep.call) call = args; args[c("resampling.method", "keep.call")] = NULL; #--------------------------------------------------------------------------------------------# resampling.method = resampling.method[1] resampling.method = match.arg(arg = resampling.method); #--------------------------------------------------------------------------------------------# out = switch(resampling.method, multi.random = do.call(what = tat_model_multirandom, args = args)) #--------------------------------------------------------------------------------------------# if(keep.call) out = c(out, list(call = call)) #--------------------------------------------------------------------------------------------# #return return(out) Train and Test